requirements.txt
oracledb>=2.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=8.0.0
pytest-cov>=5.0.0
